Move RunningInUserNS() to its own package
This allows using the utility without bringing whole of "sys" with it.
